Another big refactor
