+++ crates/supasim/src/residency.rs
@@ -3,49 +3,49 @@ use std::{collections::VecDeque, sync::Arc};
 use parking_lot::{Condvar, Mutex};
 use smallvec::SmallVec;
 
-use crate::InstanceState;
+use crate::{DEVICE_SMALLVEC_SIZE, sync::DeviceSemaphore};
 
-const DEVICE_SMALLVEC_SIZE: usize = 4;
-
-/// A semaphore with info about its device that returns to a pool on drop
-struct DeviceSemaphore<B: hal::Backend> {
-    pub inner: Option<B::Semaphore>,
-    pub device_index: usize,
-    instance: InstanceState<B>,
-}
-impl<B: hal::Backend> Drop for DeviceSemaphore<B> {
-    fn drop(&mut self) {
-        // TODO: Return the used semaphore to the instance
-    }
-}
 struct OutOfDateWait<B: hal::Backend> {
     semaphores: Vec<Arc<DeviceSemaphore<B>>>,
     needs_more_copy: bool,
 }
+#[derive(Default)]
 struct OutOfDateTracker<B: hal::Backend> {
     /// Sorted by range start. Ranges that will be out of date until their copy completes or is started.
     out_of_date_ranges: Vec<BufferAccessRange>,
     current_copies: Vec<Arc<BufferAccessFinish<B>>>,
 }
 impl<B: hal::Backend> OutOfDateTracker<B> {
-    pub fn update_range_immediate(&mut self, range: BufferAccessRange) {
+    pub fn update_range_immediate(&mut self, _range: BufferAccessRange) {
         todo!()
     }
-    pub fn invalidate_range(&mut self, range: BufferAccessRange) {
+    pub fn invalidate_range(&mut self, _range: BufferAccessRange) {
         todo!()
     }
-    pub fn update_range_delayed(&mut self, finish: BufferAccessFinish<B>) {
+    pub fn update_range_delayed(&mut self, _finish: BufferAccessFinish<B>) {
         todo!()
     }
-    pub fn get_needed_waits(&mut self, range: BufferAccessRange) -> OutOfDateWait<B> {
+    pub fn get_needed_waits(&mut self, _range: BufferAccessRange) -> OutOfDateWait<B> {
         todo!()
     }
 }
+
 /// Handles all residency and synchronizatino for a buffer
 struct DeviceResidencyState<B: hal::Backend> {
     /// The backing memory
-    buffer: Option<B::Buffer>,
+    pub buffer: Option<B::Buffer>,
     ood_tracker: OutOfDateTracker<B>,
+}
+impl<B: hal::Backend> Default for DeviceResidencyState<B> {
+    fn default() -> Self {
+        Self {
+            buffer: None,
+            ood_tracker: OutOfDateTracker {
+                out_of_date_ranges: vec![],
+                current_copies: vec![],
+            },
+        }
+    }
 }
 /// Stored in the file system
 struct StorageResidencyState<B: hal::Backend> {
@@ -70,12 +70,27 @@ struct BufferAccessFinish<B: hal::Backend> {
 }
 pub struct BufferResidency<B: hal::Backend> {
     /// Residency state for each device
-    devices: SmallVec<[DeviceResidencyState<B>; DEVICE_SMALLVEC_SIZE]>,
+    pub devices: SmallVec<[DeviceResidencyState<B>; DEVICE_SMALLVEC_SIZE]>,
     /// Residency state for the host
-    host: DeviceResidencyState<B>,
+    pub host: DeviceResidencyState<B>,
     /// Alternative to residencystate buffer for host memory
-    storage: Option<StorageResidencyState<B>>,
+    pub storage: Option<StorageResidencyState<B>>,
     /// Sorted by range start
-    read_accesses: Vec<Arc<BufferAccessFinish<B>>>,
-    write_accesses: VecDeque<Arc<BufferAccessFinish<B>>>,
+    pub read_accesses: Vec<Arc<BufferAccessFinish<B>>>,
+    pub write_accesses: VecDeque<Arc<BufferAccessFinish<B>>>,
+}
+impl<B: hal::Backend> BufferResidency<B> {
+    pub fn new(num_devices: u32) -> Self {
+        let mut devices = SmallVec::with_capacity(num_devices as usize);
+        for _ in 0..num_devices {
+            devices.push(DeviceResidencyState::default());
+        }
+        Self {
+            devices,
+            host: DeviceResidencyState::default(),
+            storage: None,
+            read_accesses: Vec::new(),
+            write_accesses: VecDeque::new(),
+        }
+    }
 }
