+++ crates/supasim/src/tests/basic_buffer_copy.rs
@@ -14,7 +14,7 @@ pub fn basic_buffer_copy<Backend: hal::Backend>(
         println!("Hello, world!");
         dev_utils::setup_trace_printer_if_env();
 
-        let instance = SupaSimInstance::<Backend>::from_hal(hal);
+        let instance = Instance::<Backend>::from_hal(hal);
         let upload_buffer = instance
             .create_buffer(&BufferDescriptor {
                 size: 16,
